show a demo page
