+++ src/components/DemoDataTable/DemoDataTable.tsx
@@ -2,21 +2,19 @@ import React from 'react'
 import styles from './DemoDataTable.module.scss'
 import { DemoData, DemoDataItem } from 'api/loopringAPI'
 
+// label, value, unit, timestamp
 type RowData = [string, string, string, number]
-
-interface DemoDataRowProps {
-  item: RowData
-}
-const DemoDataRow = ({ item }: DemoDataRowProps) => {
-  return <div />
-}
 
 interface DemoDataTableProps {
   items: RowData[]
 }
 const DemoDataTable = ({ items }: DemoDataTableProps) => {
   const func = (item: RowData, idx: number) => (
-    <DemoDataRow item={item} key={idx} />
+    <tr key={idx}>
+      <td>{item[0]}</td>
+      <td>{item[1]}</td>
+      <td>{item[2]}</td>
+    </tr>
   )
 
   return (
