+++ src/features/DemoFeature/DemoFeature.tsx
@@ -1,84 +1,25 @@
-import { createSlice, PayloadAction } from '@reduxjs/toolkit'
-import { DemoData, DemoDataItem, loadDemoData } from 'api/loopringAPI'
-import { AppThunk } from 'app/store'
+import React, { useEffect } from 'react'
+import { useSelector, useDispatch } from 'react-redux'
+import { selectDemo } from 'app/rootReducer'
+import classnames from 'classnames'
+import DemoDataTable from 'components/DemoDataTable/DemoDataTable'
+import { fetchDemoData, addDemoData } from './DemoFeatureSlice'
 
-type DemoFeatureState = DemoData & {
-  error: string | null
-  status: string | null
+interface NavbarButtonProps {
+  label: string
+  pathnames: string[]
+  currentPathname?: string
 }
 
-const initialState: DemoFeatureState = {
-  items: [],
-  error: null,
-  status: null,
+const DemoFeature = () => {
+  const dispatch = useDispatch()
+  const demo = useSelector(selectDemo)
+
+  useEffect(() => {
+    dispatch(fetchDemoData())
+  })
+
+  return <DemoDataTable />
 }
 
-const demoSlice = createSlice({
-  name: 'demo',
-  initialState,
-  reducers: {
-    loadingDemoDataStart(state, action: PayloadAction<undefined>) {
-      state.error = null
-      state.status = 'loading'
-    },
-    loadingDemoDataFailure(state, { payload }: PayloadAction<string>) {
-      state.error = payload
-      state.status = null
-    },
-    loadingDemoDataSuccess(state, { payload }: PayloadAction<DemoData>) {
-      state.error = null
-      state.status = null
-      state.items = payload.items
-    },
-    addingDemoDataStart(state, action: PayloadAction<undefined>) {
-      state.error = null
-      state.status = 'adding'
-    },
-    addingDemoDataFailure(state, { payload }: PayloadAction<string>) {
-      state.error = payload
-      state.status = null
-    },
-    addingDemoDataSuccess(state, { payload }: PayloadAction<DemoDataItem>) {
-      state.error = null
-      state.status = null
-      state.items.push(payload)
-    },
-  },
-})
-
-const {
-  loadingDemoDataStart,
-  loadingDemoDataFailure,
-  loadingDemoDataSuccess,
-  addingDemoDataStart,
-  addingDemoDataFailure,
-  addingDemoDataSuccess,
-} = demoSlice.actions
-
-export default demoSlice.reducer
-
-export const fetchDemoData = (): AppThunk => async (dispatch) => {
-  try {
-    dispatch(loadingDemoDataStart())
-
-    const [demoData] = await Promise.all([loadDemoData()])
-    dispatch(loadingDemoDataSuccess(demoData))
-  } catch (err) {
-    dispatch(loadingDemoDataFailure(err.toString()))
-  }
-}
-
-export const addDemoData = (item: DemoDataItem): AppThunk => async (
-  dispatch
-) => {
-  try {
-    dispatch(addingDemoDataStart())
-
-    // simulate a remote add API call
-    setTimeout(function () {
-      dispatch(addingDemoDataSuccess(item))
-    }, 1000)
-  } catch (err) {
-    dispatch(addingDemoDataFailure(err.toString()))
-  }
-}
+export default DemoFeature
